Model: proposals and participants include full object info
We need at least `author` in the Dict-values of `proposals`.

For convenience we now have `proposals: Dict String Proposal` (and likewise for participants).
But this implies that `id` is both in the key and the value of the Dict, which is probably an unwanted redundancy (inconsistency if `id`s differ).

So maybe we should change that later.