Add a modd config for colorized elm-make output
Elm compiler is able to produce helpfully colorized error messages.
Unfortunately these colors won't make it through webpack and elm-webpack-loader.

As a workaround we can optionally use modd alongside with webpack.

Even with modd elm-make will output color-codes only if output goes to a terminal,
not through a pipe. There are several ways to do this, see:
https://github.com/cortesi/modd/issues/26

Here we use the utility `unbuffer` from the the tool `expect`, which is available
for most Linux distros and for OSX.
https://www.nist.gov/services-resources/software/expect